air_backend/requirements.txt
Django==4.2.18
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
pymysql==1.1.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
openpyxl==3.1.2 